+++ src/config/index.js
@@ -3,4 +3,5 @@ export { default as englishContent } from 'config/languages/englishContent';
 export { default as navigationConfig } from 'config/navigationConfig';
 export { default as socialLinks } from 'config/socialLinks';
 export { default as spanishContent } from 'config/languages/spanishContent';
+export { default as supportedLanguages } from 'config/languages/supportedLanguages';
 export { defaultRedirects, routes } from 'config/routesConfig';