+++ src/config/languages/supportedLanguages.js
@@ -1,0 +1,9 @@
+import { englishContent, spanishContent } from 'config';
+
+const supportedLanguages = {
+    es: spanishContent,
+    en: englishContent,
+    default: spanishContent
+};
+
+export default supportedLanguages;